Fix for Bug 2969326 and add some debug information inside the validation of thresholds values
Inside function saHpiSensorThresholdsSet the validation of thresholds values from the caller are merged with values coming from a ThresholdsGet call. This leads to a wrong isSupport mask in the thresholds structure (reading mask instead of writing mask) and therefore wrong values are checked.
Fix: Remove the ThresholdsGet call and work with the original data from the call.